feat: dont absctract too quickly move components
